# swagger.yaml
openapi: 3.0.4
info:
  title: API de Agenda de Ajustes Pós-Cirúrgicos
  description: |
    Sistema completo para gerenciamento de agenda de ajustes de produtos pós-cirúrgicos.

    ## Características
    - Autenticação JWT com controle de permissões
    - Cadastro de pessoas com diferentes papéis
    - Gestão de agendamentos com regras de negócio
    - Upload de fotos de ajustes
    - Sistema de logs para auditoria
    - Notificações via WhatsApp
    - Relatórios e dashboard

    ## Tipos de Usuário
    - **Administrador**: Acesso total ao sistema
    - **Secretária**: Cadastro de pacientes e agendamentos
    - **Costureira**: Registro de comparecimento e upload de fotos
    - **Médico**: Visualização de relatórios
    - **Paciente**: Recebe notificações (sem acesso ao sistema)

    ## Autenticação
    A maioria dos endpoints requer autenticação via JWT. 
    Obtenha um token através do endpoint `/api/auth/login` e inclua-o no header:
    ```
    Authorization: Bearer {seu_token}
    ```

  version: 1.0.0
  contact:
    name: Ramon David Souto
    email: rdssi@hotmail.com
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: http://localhost:3000
    description: Servidor de desenvolvimento
  - url: https://api.sistema.com
    description: Servidor de produção
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido através do login
  schemas:
...
    Error:
      type: object
      properties:
        sucesso:
          type: boolean
          example: false
        mensagem:
          type: string
          example: Erro na operação
        erro:
          type: string
          example: Detalhes do erro

    # Schema de Sucesso
...
    PessoaInput:
      type: object
      required:
        - id_cidade
        - ind_tipo_pessoa
        - nom_completo
        - num_cpf
        - num_rg
        - dta_nascimento
        - des_logradouro
        - num_endereco
        - num_cep
        - nom_bairro
        - num_celular_1
        - des_email_1
      properties:
        id_cidade:
          type: integer
          description: ID da cidade
          example: 1
        ind_status:
          type: string
          enum: [ativo, inativo]
          default: ativo
          example: ativo
        ind_tipo_pessoa:
          type: string
          enum: [medico, secretaria, costureira, paciente, administrador]
          description: Tipo de pessoa/usuário
          example: paciente
        nom_completo:
          type: string
          maxLength: 255
          description: Nome completo
          example: João da Silva Santos
        num_cpf:
          type: string
          minLength: 11
          maxLength: 11
          pattern: '^\d{11}$'
          description: CPF sem formatação
          example: "12345678901"
        num_rg:
          type: string
          maxLength: 20
          description: RG
          example: "1234567"
        dta_nascimento:
          type: string
          format: date
          description: Data de nascimento
          example: "1990-05-15"
        des_logradouro:
          type: string
          maxLength: 255
          description: Logradouro (rua, avenida, etc)
          example: Rua das Flores
        num_endereco:
          type: string
          maxLength: 10
          description: Número do endereço
          example: "123"
        des_complemento:
          type: string
          maxLength: 100
          nullable: true
          description: Complemento do endereço
          example: Apartamento 201
        num_cep:
          type: string
          minLength: 8
          maxLength: 8
          pattern: '^\d{8}$'
          description: CEP sem formatação
          example: "74000000"
        nom_bairro:
          type: string
          maxLength: 100
          description: Nome do bairro/setor
          example: Centro
        num_celular_1:
          type: string
          maxLength: 15
          description: Celular principal
          example: "62999999999"
        ind_whatsapp_1:
          type: boolean
          default: true
          description: Indica se celular 1 tem WhatsApp
          example: true
        num_celular_2:
          type: string
          maxLength: 15
          nullable: true
          description: Celular secundário
          example: "62988888888"
        ind_whatsapp_2:
          type: boolean
          default: true
          description: Indica se celular 2 tem WhatsApp
          example: false
        des_email_1:
          type: string
          format: email
          maxLength: 255
          description: Email principal
          example: joao@email.com
        des_email_2:
          type: string
          format: email
          maxLength: 255
          nullable: true
          description: Email secundário
          example: joao2@email.com
        des_senha:
          type: string
          minLength: 6
          description: Senha (obrigatória para secretaria, costureira, admin e medico)
          example: senha123
    Pessoa:
      allOf:
        - $ref: "#/components/schemas/PessoaInput"
        - type: object
          properties:
            id:
              type: integer
              example: 1
            nom_cidade:
              type: string
              example: Goiânia
            nom_estado:
              type: string
              example: Goiás
            sgl_estado:
              type: string
              example: GO
            created_at:
              type: string
              format: date-time
              example: "2024-01-15T10:30:00Z"
            updated_at:
              type: string
              format: date-time
              example: "2024-01-15T10:30:00Z"
...
    Agendamento:
      type: object
      properties:
        id:
          type: integer
          example: 1
        dta_agendamento:
          type: string
          format: date-time
          example: "2024-12-31T14:00:00Z"
        ind_status:
          type: string
          enum: [agendado, confirmado, concluido, cancelado, falta]
          example: agendado
        ind_paciente_compareceu:
          type: boolean
          nullable: true
          example: null
        des_observacoes_geral:
          type: string
          nullable: true
          example: Paciente necessita de ajuste urgente
        des_observacoes_costureira:
          type: string
          nullable: true
          example: Ajuste realizado conforme solicitado
        paciente_id:
          type: integer
          example: 5
        paciente_nome:
          type: string
          example: João da Silva Santos
        paciente_cpf:
          type: string
          example: "12345678901"
        paciente_celular:
          type: string
          example: "62999999999"
        paciente_email:
          type: string
          example: joao@email.com
        costureira_id:
          type: integer
          example: 3
        costureira_nome:
          type: string
          example: Ana Costureira Santos
        secretaria_id:
          type: integer
          example: 2
        secretaria_nome:
          type: string
          example: Maria Secretária Silva
        produtos:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                example: 1
              id_produto_grade:
                type: integer
                example: 10
              nom_produto:
                type: string
                example: Blusa Manga Longa
              ref_produto:
                type: string
                example: BML-001
              num_sku:
                type: string
                example: BML-001-P-AZ
              cod_tamanho:
                type: string
                example: P
              nom_cor_estampa:
                type: string
                example: Azul
              des_ajuste_produto:
                type: string
                example: Ajustar manga esquerda
              ind_status:
                type: string
                enum: [pendente, producao, concluido, cancelado]
                example: pendente
              fotos:
                type: array
                items:
                  $ref: "#/components/schemas/Foto"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
...
    Foto:
      type: object
      properties:
        id:
          type: integer
          example: 1
        id_produto_agendamento:
          type: integer
          example: 1
        des_caminho_arquivo:
          type: string
          example: uploads/ajustes/foto-1234567890.jpg
        nom_original:
          type: string
          example: foto_ajuste.jpg
        num_tamanho_bytes:
          type: integer
          example: 2048576
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
...
    Log:
      type: object
      properties:
        id:
          type: integer
          example: 1
        id_usuario:
          type: integer
          nullable: true
          example: 1
        des_acao:
          type: string
          example: criar
        des_tabela:
          type: string
          example: tab_agendamentos
        num_registro_id:
          type: integer
          nullable: true
          example: 10
        des_detalhes:
          type: object
          example: { "paciente": "João da Silva", "data": "2024-12-31" }
        num_ip_address:
          type: string
          nullable: true
          example: "192.168.1.1"
        des_user_agent:
          type: string
          nullable: true
          example: "Mozilla/5.0..."
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        usuario_nome:
          type: string
          nullable: true
          example: Administrador do Sistema
        usuario_tipo:
          type: string
          nullable: true
          example: administrador

    # Schema de Dashboard
    Dashboard:
      type: object
      properties:
        sucesso:
          type: boolean
          example: true
        dados:
          type: object
          properties:
            mes_atual:
              type: object
              properties:
                total:
                  type: integer
                  example: 45
                concluidos:
                  type: integer
                  example: 30
                cancelados:
                  type: integer
                  example: 5
                ativos:
                  type: integer
                  example: 10
            hoje:
              type: object
              properties:
                total:
                  type: integer
                  example: 3
                agendamentos:
                  type: array
                  items:
                    $ref: "#/components/schemas/Agendamento"
            costureiras:
              type: object
              properties:
                total:
                  type: integer
                  example: 5
                disponiveis:
                  type: integer
                  example: 3
                ocupadas:
                  type: integer
                  example: 2

    # Schema de Estatísticas
...
  responses:
    UnauthorizedError:
      description: Token de autenticação ausente ou inválido
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            sucesso: false
            mensagem: Não autorizado
            erro: Token não fornecido ou inválido
    ForbiddenError:
      description: Usuário não tem permissão para executar esta ação
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            sucesso: false
            mensagem: Acesso negado
            erro: Você não tem permissão para executar esta ação
paths:
...
  /api/pessoas/costureiras/disponiveis:
    get:
      tags:
        - Pessoas
      summary: Listar costureiras com disponibilidade
      description: |
        Lista todas as costureiras indicando quais estão disponíveis para novos agendamentos.
        Uma costureira está disponível quando não possui agendamentos ativos (status agendado ou confirmado).

        **Permissões:**
        - Secretária: Pode visualizar
        - Administrador: Pode visualizar
      operationId: listarCostureirasDisponiveis
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de costureiras com status de disponibilidade
          content:
            application/json:
              schema:
                type: object
                properties:
                  sucesso:
                    type: boolean
                    example: true
                  total:
                    type: integer
                    description: Total de costureiras
                    example: 5
                  disponiveis:
                    type: integer
                    description: Número de costureiras disponíveis
                    example: 3
                  dados:
                    type: array
                    items:
                      allOf:
                        - $ref: "#/components/schemas/Pessoa"
                        - type: object
                          properties:
                            disponivel:
                              type: boolean
                              description: Indica se a costureira está disponível
                              example: true
                            agendamento_ativo:
                              nullable: true
                              allOf:
                                - $ref: "#/components/schemas/Agendamento"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
...
  /api/relatorios/dashboard:
    get:
      tags:
        - Relatórios
      summary: Dashboard geral com resumo
      description: |
        Retorna um resumo geral do sistema com estatísticas do mês atual, agendamentos de hoje e status das costureiras.

        **Permissões:**
        - Todos os usuários autenticados
      operationId: dashboard
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dados do dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /api/relatorios/logs/usuario/{id_usuario}:
    get:
      tags:
        - Relatórios
      summary: Relatório de logs por usuário
      description: |
        Lista os logs de um usuário específico.

        **Permissões:**
        - Apenas Administrador
      operationId: logsPorUsuario
      security:
        - bearerAuth: []
      parameters:
        - name: id_usuario
          in: path
          required: true
          description: ID do usuário
          schema:
            type: integer
        - name: limite
          in: query
          description: Limite de resultados
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Logs do usuário
          content:
            application/json:
              schema:
                type: object
                properties:
                  sucesso:
                    type: boolean
                    example: true
                  total:
                    type: integer
                    example: 45
                  dados:
                    type: array
                    items:
                      $ref: "#/components/schemas/Log"
        "400":
          description: ID inválido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /api/relatorios/logs/tabela/{tabela}:
    get:
      tags:
        - Relatórios
      summary: Relatório de logs por tabela
      description: |
        Lista os logs de uma tabela específica.

        **Permissões:**
        - Apenas Administrador
      operationId: logsPorTabela
      security:
        - bearerAuth: []
      parameters:
        - name: tabela
          in: path
          required: true
          description: Nome da tabela
          schema:
            type: string
        - name: registro_id
          in: query
          description: Filtrar por ID do registro
          schema:
            type: integer
      responses:
        "200":
          description: Logs da tabela
          content:
            application/json:
              schema:
                type: object
                properties:
                  sucesso:
                    type: boolean
                    example: true
                  total:
                    type: integer
                    example: 30
                  dados:
                    type: array
                    items:
                      $ref: "#/components/schemas/Log"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /api/relatorios/estatisticas/acoes:
    get:
      tags:
        - Relatórios
      summary: Estatísticas de ações
      description: |
        Retorna estatísticas sobre as ações realizadas no sistema.

        **Permissões:**
        - Apenas Administrador
      operationId: estatisticasAcoes
      security:
        - bearerAuth: []
      parameters:
        - name: data_inicio
          in: query
          description: Data inicial (formato ISO 8601)
          schema:
            type: string
            format: date-time
        - name: data_fim
          in: query
          description: Data final (formato ISO 8601)
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Estatísticas de ações
          content:
            application/json:
              schema:
                type: object
                properties:
                  sucesso:
                    type: boolean
                    example: true
                  periodo:
                    type: object
                    properties:
                      data_inicio:
                        type: string
                        format: date-time
                      data_fim:
                        type: string
                        format: date-time
                  dados:
                    type: array
                    items:
                      type: object
                      properties:
                        des_acao:
                          type: string
                          example: criar
                        total:
                          type: integer
                          example: 150
                        usuarios_unicos:
                          type: integer
                          example: 5
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /api/relatorios/estatisticas/usuarios:
    get:
      tags:
        - Relatórios
      summary: Estatísticas de usuários
      description: |
        Retorna estatísticas sobre a atividade dos usuários.

        **Permissões:**
        - Apenas Administrador
      operationId: estatisticasUsuarios
      security:
        - bearerAuth: []
      parameters:
        - name: data_inicio
          in: query
          description: Data inicial (formato ISO 8601)
          schema:
            type: string
            format: date-time
        - name: data_fim
          in: query
          description: Data final (formato ISO 8601)
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Estatísticas de usuários
          content:
            application/json:
              schema:
                type: object
                properties:
                  sucesso:
                    type: boolean
                    example: true
                  periodo:
                    type: object
                    properties:
                      data_inicio:
                        type: string
                        format: date-time
                      data_fim:
                        type: string
                        format: date-time
                  dados:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        nom_completo:
                          type: string
                          example: Administrador do Sistema
                        ind_tipo_pessoa:
                          type: string
                          example: administrador
                        total_acoes:
                          type: integer
                          example: 200
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
...
  /api/relatorios/produtividade/costureiras:
    get:
      tags:
        - Relatórios
      summary: Relatório de produtividade por costureira
      description: |
        Gera relatório de produtividade das costureiras.

        **Permissões:**
        - Apenas Administrador
      operationId: produtividadeCostureiras
      security:
        - bearerAuth: []
      parameters:
        - name: data_inicio
          in: query
          description: Data inicial (formato ISO 8601)
          schema:
            type: string
            format: date-time
        - name: data_fim
          in: query
          description: Data final (formato ISO 8601)
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Relatório de produtividade
          content:
            application/json:
              schema:
                type: object
                properties:
                  sucesso:
                    type: boolean
                    example: true
                  periodo:
                    type: object
                    properties:
                      data_inicio:
                        type: string
                        format: date-time
                      data_fim:
                        type: string
                        format: date-time
                  total_costureiras:
                    type: integer
                    example: 5
                  dados:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 3
                        nome:
                          type: string
                          example: Ana Costureira Santos
                        total_agendamentos:
                          type: integer
                          example: 45
                        concluidos:
                          type: integer
                          example: 40
                        cancelados:
                          type: integer
                          example: 2
                        faltas:
                          type: integer
                          example: 3
                        em_andamento:
                          type: integer
                          example: 0
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"

  # ==================== HEALTH CHECK ====================
...